refactor(launcher): decompose settings dialog
