+++ AGENT_HANDOFF.md
@@ -5,6 +5,29 @@ Update this file with every PR and every push to main.
 
 ## Where the Repo Is Heading
 
+- **Launcher-settings size slice (local, no publish)** — branch
+  `codex/settings-dialog-size-decomposition` starts at exact draft PR #8486 head
+  `624043537a5ab10aa7ef56dc61685a004b872c0c`. It extracts runtime dependency
+  probes and the WSL setup dialog into `settings_runtime.py`, and extracts
+  diagnostics, log synchronization, and process-management behavior into the
+  private `_settings_auxiliary_tabs.py` mixin. `SettingsWidget` preserves its
+  constructor, signal, tab constants, control attributes, historical methods,
+  and runtime compatibility imports. `settings_dialog.py` is now 1,124 lines
+  (down from 2,190); its file-size and expired module-size exceptions plus the
+  obsolete `WslScriptDialog._setup_ui` long-function exception are removed
+  without renewal. All 44 focused settings/launcher contract tests pass, as do
+  changed-file Ruff/format/compile, architecture, file-size, and error-ratchet
+  gates. A wider 166-test launcher selection is 148 passed and 18 inherited
+  failures caused by shared theme/UI export drift; the exact parent produces
+  the same 18 failures (145 passed without the three new decomposition tests).
+  The unrestricted launcher suite also reaches the same Windows access
+  violation in `test_run_launcher` on the exact parent. The official Python
+  3.12 MyPy wrapper excludes all three launcher modules by repository policy.
+  The only remaining global module-size failure is the parent-identical expired
+  `launcher_ui_setup.py` exception. The cited
+  #5922 and #7341/#7342 issues are closed or unrelated, so accurate open issue
+  tracking remains a publication prerequisite; do not claim issue completion.
+
 - **#8485 Simscape 3D-viewer size slice — draft PR
   [#8486](https://github.com/D-sorganization/UpstreamDrift/pull/8486)** — branch
   `fix/viewer-3d-module-size-decomposition` is stacked on draft PR #8484 at
