+++ src/launchers/settings_dialog.py
@@ -9,14 +9,11 @@ from __future__ import annotations
 
 import functools
 import os
-import subprocess
 import time
 from collections.abc import Callable
-from dataclasses import dataclass
-from pathlib import Path
 from typing import Any
 
-from PyQt6.QtCore import QThread, Qt, pyqtSignal
+from PyQt6.QtCore import Qt, pyqtSignal
 from PyQt6.QtWidgets import (
     QCheckBox,
     QComboBox,
@@ -40,7 +37,16 @@ from src.launchers.build_close_guard import confirm_cancel_running_build_for_clo
 from src.launchers.docker_manager import DockerBuildThread
 from src.launchers.docker_profile_info import load_docker_profiles
 from src.launchers.launcher_constants import DOCKER_STAGES
-from src.launchers.hover_copy_browser import HoverCopyTextBrowser
+from src.launchers._settings_auxiliary_tabs import SettingsAuxiliaryTabsMixin
+from src.launchers.settings_runtime import (
+    RuntimeDependencyCheckFailure as RuntimeDependencyCheckFailure,
+    RuntimeDependencyCheckWorker,
+    RuntimeDependencyReport,
+    WslScriptDialog,
+    check_docker_dependencies_report as _check_docker_dependencies_report,
+    check_wsl_dependencies_report as _check_wsl_dependencies_report,
+    compare_version_strings as _compare_version_strings,
+)
 from src.shared.python.docker_config import DOCKER_IMAGE_ENGINE as DOCKER_IMAGE_NAME
 from src.shared.python.logging_pkg.logging_config import get_logger
 from src.shared.python.theme.style_constants import Styles
@@ -58,268 +64,6 @@ TAB_STARTUP = 5
 TAB_NOTIFICATIONS = 6
 TAB_PERFORMANCE = 7
 TAB_PROCESSES = 8
-
-_RUNTIME_DEPENDENCIES = {
-    "numpy": ">=1.26.4",
-    "scipy": ">=1.13.1",
-    "mujoco": ">=3.6.0",
-    "pydrake": ">=1.22.0",
-    "pinocchio": ">=2.6.0",
-    "opensim": ">=4.4.0",
-    "myosuite": ">=2.0.0",
-}
-_RUNTIME_DISPLAY_NAMES = {
-    "numpy": "NumPy",
-    "scipy": "SciPy",
-    "mujoco": "MuJoCo",
-    "pydrake": "Drake (PyDrake)",
-    "pinocchio": "Pinocchio",
-    "opensim": "OpenSim",
-    "myosuite": "MyoSuite",
-}
-_RUNTIME_OPTIONAL_DEPS = {"pydrake", "pinocchio", "opensim", "myosuite"}
-_DEPENDENCY_PROBE_SCRIPT = (
-    "import importlib.metadata\n"
-    "import sys\n"
-    "res = []\n"
-    f"for p in {list(_RUNTIME_DEPENDENCIES)!r}:\n"
-    "    try:\n"
-    "        __import__(p)\n"
-    "        try:\n"
-    "            v = importlib.metadata.version(p)\n"
-    "        except Exception:\n"
-    "            v = getattr(sys.modules[p], '__version__', 'Unknown')\n"
-    "        res.append(f'{p}:{v}')\n"
-    "    except ImportError:\n"
-    "        res.append(f'{p}:Missing')\n"
-    "print(','.join(res))"
-)
-
-
-@dataclass(frozen=True)
-class RuntimeDependencyReport:
-    """Structured result from a background runtime dependency check."""
-
-    dialog_title: str
-    table_title: str
-    environment_name: str
-    check_results: list[dict[str, Any]]
-
-
-class RuntimeDependencyCheckFailure(RuntimeError):
-    """A dependency check completed with a user-facing warning or error."""
-
-    def __init__(self, severity: str, title: str, html: str) -> None:
-        super().__init__(html)
-        self.severity = severity
-        self.title = title
-        self.html = html
-
-
-class RuntimeDependencyCheckWorker(QThread):
-    """Run Docker or WSL dependency checks without blocking the GUI thread."""
-
-    succeeded = pyqtSignal(object)
-    failed = pyqtSignal(str, str, str)
-
-    def __init__(
-        self,
-        check_fn: Callable[[], RuntimeDependencyReport],
-        parent: QWidget | None = None,
-    ) -> None:
-        super().__init__(parent)
-        self._check_fn = check_fn
-
-    def run(self) -> None:
-        try:
-            self.succeeded.emit(self._check_fn())
-        except RuntimeDependencyCheckFailure as exc:
-            self.failed.emit(exc.severity, exc.title, exc.html)
-        except (subprocess.SubprocessError, FileNotFoundError, OSError) as exc:
-            self.failed.emit(
-                "critical",
-                "Dependency Check",
-                f"<h3>Runtime Dependency Check Failed</h3><p>{exc}</p>",
-            )
-
-
-def _compare_version_strings(installed: str, required_spec: str) -> bool:
-    """Compare installed version string with required specification."""
-    if installed == "Unknown":
-        return True
-    if installed == "Missing":
-        return False
-
-    if required_spec.startswith(">="):
-        op = ">="
-        req_ver = required_spec[2:]
-    elif required_spec.startswith("=="):
-        op = "=="
-        req_ver = required_spec[2:]
-    else:
-        op = ">="
-        req_ver = required_spec
-
-    try:
-
-        def to_int_list(v_str: str) -> list[int]:
-            parts = []
-            for p in v_str.split("."):
-                digits = []
-                for c in p:
-                    if c.isdigit():
-                        digits.append(c)
-                    else:
-                        break
-                parts.append(int("".join(digits)) if digits else 0)
-            return parts
-
-        inst_parts = to_int_list(installed)
-        req_parts = to_int_list(req_ver)
-        length = max(len(inst_parts), len(req_parts))
-        inst_parts += [0] * (length - len(inst_parts))
-        req_parts += [0] * (length - len(req_parts))
-
-        if op == ">=":
-            return inst_parts >= req_parts
-        if op == "==":
-            return inst_parts == req_parts
-    except (ValueError, TypeError, IndexError):
-        pass
-    return True
-
-
-def _parse_dependency_output(output: str) -> dict[str, str]:
-    parsed_deps = {}
-    for item in output.split(","):
-        if ":" in item:
-            key, value = item.split(":", 1)
-            parsed_deps[key] = value
-    return parsed_deps
-
-
-def _build_dependency_check_results(
-    parsed_deps: dict[str, str],
-) -> list[dict[str, Any]]:
-    check_results = []
-    for package_name, requirement in _RUNTIME_DEPENDENCIES.items():
-        installed_version = parsed_deps.get(package_name, "Missing")
-        if installed_version == "Missing":
-            status = "warn" if package_name in _RUNTIME_OPTIONAL_DEPS else "error"
-        else:
-            status = (
-                "ok"
-                if _compare_version_strings(installed_version, requirement)
-                else "error"
-            )
-        check_results.append(
-            {
-                "name": _RUNTIME_DISPLAY_NAMES[package_name],
-                "required": requirement,
-                "installed": installed_version,
-                "status": status,
-            }
-        )
-    return check_results
-
-
-def _check_docker_dependencies_report() -> RuntimeDependencyReport:
-    from src.launchers.docker_manager import get_docker_cmd
-
-    cmd = get_docker_cmd()
-    inspect_cmd = cmd + ["image", "inspect", "upstream-drift:engine"]
-    try:
-        res = subprocess.run(inspect_cmd, capture_output=True, text=True, timeout=10.0)
-    except subprocess.TimeoutExpired as exc:
-        msg = (
-            "<h3>Docker Connection Timeout</h3>"
-            "<p>Failed to communicate with the Docker daemon within the timeout limit (10.0s):</p>"
-            f"<pre style='color:#f87171;'>{exc}</pre>"
-            "<p>This typically occurs if the WSL2 subsystem or Docker Desktop is starting up, frozen, or not running.</p>"
-            "<p>Please verify that:</p>"
-            "<ol>"
-            "  <li><b>Docker Desktop</b> is running.</li>"
-            "  <li><b>WSL integration</b> is enabled for your distribution in Docker Desktop settings.</li>"
-            "  <li>You can run <code>wsl docker ps</code> in a terminal without hanging.</li>"
-            "</ol>"
-        )
-        raise RuntimeDependencyCheckFailure(
-            "critical", "Docker Dependency Check", msg
-        ) from exc
-
-    if res.returncode != 0:
-        msg = (
-            "<h3>Docker Environment Status: Missing Image</h3>"
-            "<p>The <b>upstream-drift:engine</b> image has not been built yet.</p>"
-            "<p>Please click the <b>Build Image</b> button below to build the container first.</p>"
-        )
-        raise RuntimeDependencyCheckFailure("warning", "Docker Dependency Check", msg)
-
-    run_cmd = cmd + [
-        "run",
-        "--rm",
-        "upstream-drift:engine",
-        "python3",
-        "-c",
-        _DEPENDENCY_PROBE_SCRIPT,
-    ]
-    try:
-        res_run = subprocess.run(run_cmd, capture_output=True, text=True, timeout=60.0)
-    except subprocess.TimeoutExpired as exc:
-        msg = (
-            "<h3>Docker Environment Status: Timeout</h3>"
-            "<p>Docker container dependency check timed out.</p>"
-            "<p>The Docker image exists, but a cold container start did not finish within 60 seconds.</p>"
-        )
-        raise RuntimeDependencyCheckFailure(
-            "warning", "Docker Dependency Check", msg
-        ) from exc
-
-    if res_run.returncode != 0:
-        msg = (
-            "<h3>Docker Environment Status: Degraded</h3>"
-            f"<p>The image is built, but running probe failed (Exit Code {res_run.returncode}).</p>"
-            f"<pre style='color:#f87171;'>{res_run.stderr.strip()}</pre>"
-            "<p>Verify Docker is running and has access to execute containers in WSL.</p>"
-        )
-        raise RuntimeDependencyCheckFailure("warning", "Docker Dependency Check", msg)
-
-    return RuntimeDependencyReport(
-        dialog_title="Docker Dependency Check",
-        table_title="Docker Container Environment",
-        environment_name="Docker Container",
-        check_results=_build_dependency_check_results(
-            _parse_dependency_output(res_run.stdout.strip())
-        ),
-    )
-
-
-def _check_wsl_dependencies_report() -> RuntimeDependencyReport:
-    from src.shared.python.data_io.path_utils import get_repo_root
-
-    repo_root = get_repo_root()
-    venv_path = repo_root / ".venv-wsl"
-    python_exec = "./.venv-wsl/bin/python" if venv_path.exists() else "python3"
-    cmd = ["wsl", python_exec, "-c", _DEPENDENCY_PROBE_SCRIPT]
-    res = subprocess.run(cmd, capture_output=True, text=True, timeout=5.0)
-    if res.returncode != 0:
-        msg = (
-            "<h3>WSL Environment Status: Not Set Up</h3>"
-            "<p>WSL environment could not be checked or has not been initialized.</p>"
-            f"<p>Details: Exit Code {res.returncode}</p>"
-            f"<pre style='color:#f87171;'>{res.stderr.strip()}</pre>"
-            "<p>Please click the <b>WSL Setup</b> button to view the setup script and run it in WSL.</p>"
-        )
-        raise RuntimeDependencyCheckFailure("warning", "WSL Dependency Check", msg)
-
-    return RuntimeDependencyReport(
-        dialog_title="WSL Dependency Check",
-        table_title="WSL2 Environment Status",
-        environment_name=f"WSL ({python_exec})",
-        check_results=_build_dependency_check_results(
-            _parse_dependency_output(res.stdout.strip())
-        ),
-    )
 
 
 def validate_tab_index(tab_index: int) -> int:
@@ -342,199 +86,7 @@ def validate_tab_index(tab_index: int) -> int:
     return tab_index
 
 
-class WslScriptDialog(QDialog):
-    """Dialog to inspect, copy, and run the WSL dependencies installation script."""
-
-    def __init__(self, parent: QWidget | None = None) -> None:
-        super().__init__(parent)
-        self.setWindowTitle("WSL Dependency Setup Script")
-        self.resize(700, 550)
-        self._setup_ui()
-
-    def _setup_ui(self) -> None:
-        layout = QVBoxLayout(self)
-
-        # Header description
-        info_label = QLabel(
-            "<h3>WSL2 Dependency Setup</h3>"
-            "<p>This script installs all system and Python dependencies needed to run "
-            "UpstreamDrift physics engines (including Drake, Pinocchio, MuJoCo, and OpenSim) inside WSL2 Ubuntu.</p>"
-        )
-        info_label.setWordWrap(True)
-        info_label.setTextFormat(Qt.TextFormat.RichText)
-        layout.addWidget(info_label)
-
-        # Commands row
-        cmd_group = QGroupBox("Run Commands")
-        cmd_layout = QVBoxLayout(cmd_group)
-
-        # 1. WSL shell command
-        wsl_cmd_row = QHBoxLayout()
-        wsl_cmd_row.addWidget(QLabel("<b>Inside WSL:</b>"))
-        self.lbl_wsl_cmd = QLabel(
-            "<code>bash scripts/install_wsl_dependencies.sh</code>"
-        )
-        self.lbl_wsl_cmd.setTextInteractionFlags(
-            Qt.TextInteractionFlag.TextSelectableByMouse
-        )
-        btn_copy_wsl_cmd = QPushButton("Copy")
-        btn_copy_wsl_cmd.setFixedWidth(60)
-        btn_copy_wsl_cmd.clicked.connect(
-            lambda: self._copy_text(
-                "bash scripts/install_wsl_dependencies.sh", "WSL shell command"
-            )
-        )
-        wsl_cmd_row.addWidget(self.lbl_wsl_cmd)
-        wsl_cmd_row.addStretch()
-        wsl_cmd_row.addWidget(btn_copy_wsl_cmd)
-        cmd_layout.addLayout(wsl_cmd_row)
-
-        # 2. Windows cmd command
-        win_cmd_row = QHBoxLayout()
-        win_cmd_row.addWidget(QLabel("<b>From Windows:</b>"))
-        self.lbl_win_cmd = QLabel(
-            "<code>wsl bash scripts/install_wsl_dependencies.sh</code>"
-        )
-        self.lbl_win_cmd.setTextInteractionFlags(
-            Qt.TextInteractionFlag.TextSelectableByMouse
-        )
-        btn_copy_win_cmd = QPushButton("Copy")
-        btn_copy_win_cmd.setFixedWidth(60)
-        btn_copy_win_cmd.clicked.connect(
-            lambda: self._copy_text(
-                "wsl bash scripts/install_wsl_dependencies.sh", "Windows run command"
-            )
-        )
-        win_cmd_row.addWidget(self.lbl_win_cmd)
-        win_cmd_row.addStretch()
-        win_cmd_row.addWidget(btn_copy_win_cmd)
-        cmd_layout.addLayout(win_cmd_row)
-
-        layout.addWidget(cmd_group)
-
-        # Interactive Run button
-        btn_run_row = QHBoxLayout()
-        self.btn_run_terminal = QPushButton(
-            "Run script in interactive WSL Terminal window"
-        )
-        self.btn_run_terminal.setToolTip(
-            "Open a new Windows Command Prompt that runs the script inside WSL. "
-            "Recommended if sudo password prompt is required."
-        )
-        self.btn_run_terminal.clicked.connect(self._run_in_terminal)
-        self.btn_run_terminal.setStyleSheet("""
-            QPushButton {
-                background-color: #0366d6;
-                color: white;
-                font-weight: bold;
-                padding: 8px;
-                border-radius: 4px;
-            }
-            QPushButton:hover {
-                background-color: #0056b3;
-            }
-        """)
-        btn_run_row.addWidget(self.btn_run_terminal)
-        layout.addLayout(btn_run_row)
-
-        # Script content viewer
-        content_group = QGroupBox(
-            "Script Contents (scripts/install_wsl_dependencies.sh)"
-        )
-        content_layout = QVBoxLayout(content_group)
-
-        self.txt_content = QTextEdit()
-        self.txt_content.setReadOnly(True)
-        # Set dark-ish monospace font style
-        self.txt_content.setStyleSheet(
-            "font-family: 'Courier New', monospace; background-color: #1e1e1e; color: #d4d4d4;"
-        )
-
-        # Load script content
-        script_text = self._load_script_content()
-        self.txt_content.setPlainText(script_text)
-        content_layout.addWidget(self.txt_content)
-
-        # Action buttons for content
-        content_btn_row = QHBoxLayout()
-        btn_copy_content = QPushButton("Copy Script Contents")
-        btn_copy_content.clicked.connect(
-            lambda: self._copy_text(self.txt_content.toPlainText(), "Script contents")
-        )
-        content_btn_row.addWidget(btn_copy_content)
-        content_btn_row.addStretch()
-        content_layout.addLayout(content_btn_row)
-
-        layout.addWidget(content_group)
-
-        # Dialog buttons
-        dialog_buttons = QDialogButtonBox(QDialogButtonBox.StandardButton.Close)
-        dialog_buttons.rejected.connect(self.reject)
-        layout.addWidget(dialog_buttons)
-
-    def _load_script_content(self) -> str:
-        from src.shared.python.data_io.path_utils import get_repo_root
-
-        script_path = get_repo_root() / "scripts" / "install_wsl_dependencies.sh"
-        if script_path.exists():
-            try:
-                return script_path.read_text(encoding="utf-8")
-            except OSError as e:
-                return f"# Error loading script contents: {e}"
-        return "# Error: scripts/install_wsl_dependencies.sh not found."
-
-    def _copy_text(self, text: str, name: str) -> None:
-        from PyQt6.QtGui import QGuiApplication
-
-        clipboard = QGuiApplication.clipboard()
-        if clipboard:
-            clipboard.setText(text)
-            from PyQt6.QtWidgets import QMessageBox
-
-            QMessageBox.information(
-                self, "Copied", f"<p>{name} has been copied to the clipboard.</p>"
-            )
-
-    def _run_in_terminal(self) -> None:
-        import subprocess
-
-        # Use cmd.exe start to run in a separate persistent command prompt window
-        from src.shared.python.data_io.path_utils import get_repo_root
-
-        repo_root = get_repo_root()
-
-        try:
-            subprocess.Popen(
-                [
-                    "cmd.exe",
-                    "/c",
-                    "start",
-                    "cmd.exe",
-                    "/k",
-                    "wsl bash scripts/install_wsl_dependencies.sh",
-                ],
-                cwd=str(repo_root),
-            )
-            from PyQt6.QtWidgets import QMessageBox
-
-            QMessageBox.information(
-                self,
-                "WSL Setup Script Started",
-                "<p>Launched the WSL setup script in a new terminal window.</p>"
-                "<p>Please check the opened window to monitor progress, enter your password if prompted, "
-                "and verify installation success.</p>",
-            )
-        except OSError as e:
-            from PyQt6.QtWidgets import QMessageBox
-
-            QMessageBox.critical(
-                self,
-                "Execution Error",
-                f"<p>Failed to launch terminal process: {e}</p>",
-            )
-
-
-class SettingsWidget(QWidget):
+class SettingsWidget(SettingsAuxiliaryTabsMixin, QWidget):
     """Settings widget with Layout, Configuration, Diagnostics, MCP Servers, and Preferences tabs.
 
     Tab order:
@@ -1100,70 +652,6 @@ class SettingsWidget(QWidget):
 
         return scroll
 
-    # ── Diagnostics tab ─────────────────────────────────────────────
-
-    def _create_diagnostics_tab(self) -> QWidget:
-        """Diagnostics tab: system checks, error log viewer, terminal output."""
-        tab = QWidget()
-        tab_layout = QVBoxLayout(tab)
-
-        # System checks browser
-        self._diag_browser = HoverCopyTextBrowser()
-        self._diag_browser.setOpenExternalLinks(False)
-        self._diag_browser.setStyleSheet(Styles.CONSOLE_DIAGNOSTICS)
-        tab_layout.addWidget(self._diag_browser, stretch=3)
-
-        if self._diagnostics_data:
-            self._render_diagnostics(self._diagnostics_data)
-
-        # Process output log viewer
-        proc_group = QGroupBox("Process Output Log (recent)")
-        proc_inner = QVBoxLayout(proc_group)
-        self._proc_log_viewer = QTextEdit()
-        self._proc_log_viewer.setReadOnly(True)
-        self._proc_log_viewer.setMaximumHeight(180)
-        self._proc_log_viewer.setStyleSheet(Styles.CONSOLE_LOG_GREEN)
-        proc_inner.addWidget(self._proc_log_viewer)
-        tab_layout.addWidget(proc_group, stretch=1)
-
-        # Application log viewer
-        log_group = QGroupBox("Application Log (recent)")
-        log_inner = QVBoxLayout(log_group)
-        self._log_viewer = QTextEdit()
-        self._log_viewer.setReadOnly(True)
-        self._log_viewer.setMaximumHeight(160)
-        self._log_viewer.setStyleSheet(Styles.CONSOLE_LOG_LIGHT)
-        log_inner.addWidget(self._log_viewer)
-        tab_layout.addWidget(log_group, stretch=1)
-
-        # Load recent log lines
-        self._load_process_log()
-        self._load_app_log()
-
-        # Action buttons
-        btn_row = QHBoxLayout()
-        btn_row.addStretch()
-
-        self.btn_sync_tools = QPushButton("Sync Shared Tools")
-        self.btn_sync_tools.setToolTip(
-            "Synchronize git submodules and fetch latest updates"
-        )
-        self.btn_sync_tools.clicked.connect(self._sync_shared_tools)
-        btn_row.addWidget(self.btn_sync_tools)
-
-        btn_refresh = QPushButton("Re-run Diagnostics")
-        btn_refresh.setToolTip("Run all diagnostic checks again")
-        btn_refresh.clicked.connect(self._refresh_diagnostics)
-        btn_row.addWidget(btn_refresh)
-
-        btn_refresh_log = QPushButton("Refresh Logs")
-        btn_refresh_log.setToolTip("Reload all log files")
-        btn_refresh_log.clicked.connect(self._refresh_all_logs)
-        btn_row.addWidget(btn_refresh_log)
-
-        tab_layout.addLayout(btn_row)
-        return tab
-
     # ── MCP Servers tab ─────────────────────────────────────────────
 
     def _create_mcp_servers_tab(self) -> QWidget:
@@ -1196,210 +684,6 @@ class SettingsWidget(QWidget):
             "MCP server configuration changed.\n\n"
             "Restart the Sidekick chat session for changes to take effect.",
         )
-
-    def _load_app_log(self) -> None:
-        """Load recent lines from the application log file."""
-        log_candidates = [
-            Path.cwd() / "app_launch.log",
-            Path.home() / ".golf_modeling_suite" / "launcher.log",
-        ]
-        for log_path in log_candidates:
-            if log_path.exists():
-                try:
-                    text = log_path.read_text(encoding="utf-8", errors="replace")
-                    lines = text.strip().splitlines()
-                    recent = "\n".join(lines[-200:])
-                    self._log_viewer.setPlainText(recent)
-                    from PyQt6.QtGui import QTextCursor
-
-                    self._log_viewer.moveCursor(QTextCursor.MoveOperation.End)
-                    return
-                except (RuntimeError, ValueError, AttributeError, OSError) as e:
-                    logger.debug("Could not display log file %s: %s", log_path, e)
-        self._log_viewer.setPlainText("(No log file found)")
-
-    def _load_process_log(self) -> None:
-        """Load recent lines from the process output log file."""
-        log_path = Path.home() / ".golf_modeling_suite" / "process_output.log"
-        if log_path.exists():
-            try:
-                text = log_path.read_text(encoding="utf-8", errors="replace")
-                lines = text.strip().splitlines()
-                recent = "\n".join(lines[-300:])
-                self._proc_log_viewer.setPlainText(recent)
-                from PyQt6.QtGui import QTextCursor
-
-                self._proc_log_viewer.moveCursor(QTextCursor.MoveOperation.End)
-                return
-            except (RuntimeError, ValueError, AttributeError, OSError) as e:
-                logger.debug("Could not display process log %s: %s", log_path, e)
-        self._proc_log_viewer.setPlainText(
-            "(No process output log yet — launch a model to generate output)"
-        )
-
-    def _refresh_all_logs(self) -> None:
-        """Refresh both log viewers."""
-        self._load_process_log()
-        self._load_app_log()
-
-    def _render_diagnostics(self, data: dict[str, Any]) -> None:
-        """Render diagnostics results as styled HTML."""
-        if data is None:
-            raise ValueError("data must be provided")
-        summary = data.get("summary", {})
-        checks = data.get("checks", [])
-        runtime = data.get("runtime_state", {})
-        recommendations = data.get("recommendations", [])
-
-        html = self._render_diag_summary(summary)
-        html += self._render_diag_checks(checks)
-        html += self._render_diag_engines(checks)
-        html += self._render_diag_runtime(runtime)
-        html += self._render_diag_recommendations(recommendations)
-
-        self._diag_browser.setHtml(html)
-
-    def _render_diag_summary(self, summary: dict) -> str:
-        if summary is None:
-            raise ValueError("summary must be provided")
-        status = summary.get("status", "unknown").upper()
-        passed = summary.get("passed", 0)
-        failed = summary.get("failed", 0)
-        warnings = summary.get("warnings", 0)
-        total = summary.get("total_checks", passed + failed + warnings)
-
-        status_color = "#2da44e" if status == "HEALTHY" else "#d29922"
-        return f"""
-        <div style="margin-bottom: 12px;">
-            <h2 style="color:{status_color}; margin: 0;">Status: {status}</h2>
-            <p><b>{total} checks:</b>
-                <span style="color:#2da44e;">{passed} passed</span>,
-                <span style="color:#f85149;">{failed} failed</span>,
-                <span style="color:#d29922;">{warnings} warnings</span>
-            </p>
-        </div>
-        """
-
-    def _render_diag_checks(self, checks: list) -> str:
-        if checks is None:
-            raise ValueError("checks must be provided")
-        html = "<h3>Check Results</h3><table style='width:100%;'>"
-        for check in checks:
-            icon = {"pass": "&#9989;", "fail": "&#10060;", "warning": "&#9888;"}.get(
-                check["status"], "&#8226;"
-            )
-            color = {"pass": "#2da44e", "fail": "#f85149", "warning": "#d29922"}.get(
-                check["status"], "#d4d4d4"
-            )
-            duration = check.get("duration_ms", 0)
-            html += (
-                f"<tr><td style='color:{color}; padding:2px 6px;'>{icon}</td>"
-                f"<td style='padding:2px 6px;'><b>{check['name']}</b></td>"
-                f"<td style='padding:2px 6px; color:#a0a0a0;'>{check['message']}</td>"
-                f"<td style='padding:2px 6px; color:#666;'>{duration:.0f}ms</td></tr>"
-            )
-        html += "</table>"
-        return html
-
-    def _render_diag_engines(self, checks: list) -> str:
-        if checks is None:
-            raise ValueError("checks must be provided")
-        engine_check = next(
-            (c for c in checks if c["name"] == "engine_availability"), None
-        )
-        engines = (
-            engine_check.get("details", {}).get("engines", []) if engine_check else []
-        )
-        if not engines:
-            return ""
-
-        html = "<h3>Physics Engines</h3>"
-        html += (
-            "<table style='width:100%; border-collapse:collapse;'>"
-            "<tr style='border-bottom:1px solid #333;'>"
-            "<th style='padding:4px 8px; text-align:left;'>Engine</th>"
-            "<th style='padding:4px 8px; text-align:left;'>Status</th>"
-            "<th style='padding:4px 8px; text-align:left;'>Version</th>"
-            "<th style='padding:4px 8px; text-align:left;'>Details</th>"
-            "</tr>"
-        )
-        for eng in engines:
-            installed = eng.get("installed", False)
-            icon = "&#9989;" if installed else "&#10060;"
-            color = "#2da44e" if installed else "#f85149"
-            name = eng.get("name", "?").replace("_", " ").title()
-            version = eng.get("version") or "-"
-            diag = eng.get("diagnostic", "")
-            missing = eng.get("missing_deps", [])
-            detail_str = diag
-            if missing and not installed:
-                detail_str = f"Missing: {', '.join(missing[:3])}"
-            html += (
-                f"<tr>"
-                f"<td style='padding:3px 8px;'><b>{name}</b></td>"
-                f"<td style='padding:3px 8px; color:{color};'>{icon} "
-                f"{'Installed' if installed else 'Not installed'}</td>"
-                f"<td style='padding:3px 8px; color:#a0a0a0;'>{version}</td>"
-                f"<td style='padding:3px 8px; color:#888;'>{detail_str}</td>"
-                f"</tr>"
-            )
-        html += "</table>"
-        return html
-
-    def _render_diag_runtime(self, runtime: dict) -> str:
-        if runtime is None:
-            raise ValueError("runtime must be provided")
-        if not runtime:
-            return ""
-        html = "<h3>Runtime State</h3><ul>"
-        html += (
-            f"<li>Available models: {runtime.get('available_models_count', '?')}</li>"
-        )
-        html += f"<li>Tile order: {runtime.get('model_order_count', '?')}</li>"
-        html += f"<li>Model cards: {runtime.get('model_cards_count', '?')}</li>"
-        html += f"<li>Registry loaded: {runtime.get('registry_loaded', '?')}</li>"
-        html += f"<li>Docker available: {runtime.get('docker_available', '?')}</li>"
-        html += "</ul>"
-        return html
-
-    def _render_diag_recommendations(self, recommendations: list) -> str:
-        if recommendations is None:
-            raise ValueError("recommendations must be provided")
-        if not recommendations:
-            return ""
-        html = "<h3>Recommendations</h3><ul>"
-        for rec in recommendations[:8]:
-            html += f"<li>{rec}</li>"
-        html += "</ul>"
-        return html
-
-    def _refresh_diagnostics(self) -> None:
-        """Re-run diagnostics and update the display."""
-        try:
-            from src.launchers.launcher_diagnostics import LauncherDiagnostics
-
-            diag = LauncherDiagnostics()
-            results = diag.run_all_checks()
-
-            launcher = self._launcher
-            if launcher and hasattr(launcher, "available_models"):
-                results["runtime_state"] = {
-                    "available_models_count": len(launcher.available_models),
-                    "available_model_ids": list(launcher.available_models.keys()),
-                    "model_order_count": len(launcher.model_order),
-                    "model_order": launcher.model_order,
-                    "model_cards_count": len(launcher.model_cards),
-                    "selected_model": launcher.selected_model,
-                    "docker_available": launcher.docker_available,
-                    "registry_loaded": launcher.registry is not None,
-                }
-
-            self._diagnostics_data = results
-            self._render_diagnostics(results)
-        except ImportError as e:
-            self._diag_browser.setHtml(
-                f"<p style='color:#f85149;'>Error running diagnostics: {e}</p>"
-            )
 
     def _on_reset_layout(self) -> None:
         self.reset_layout_requested.emit()
@@ -1555,203 +839,6 @@ class SettingsWidget(QWidget):
         if hasattr(self, "_build_start_time"):
             elapsed = time.monotonic() - self._build_start_time
             self._build_status.setText(f"Building... ({elapsed:.0f}s elapsed)")
-
-    def _on_tab_changed(self, index: int) -> None:
-        """Handle tab change for lazy diagnostics loading and processes refresh.
-
-        When the Diagnostics tab is first selected, run diagnostics
-        asynchronously and populate the display. Subsequent selections
-        do not re-run diagnostics.
-
-        Args:
-            index: The new tab index.
-        """
-        if index == TAB_PROCESSES:
-            self.refresh_processes_ui()
-            return
-
-        if index != TAB_DIAGNOSTICS or self._diagnostics_loaded:
-            return
-
-        # Mark as loaded to prevent re-running on subsequent tab visits
-        self._diagnostics_loaded = True
-
-        # Show loading indicator
-        self._diag_browser.setHtml(
-            "<p style='color:#d29922;'>Running diagnostics...</p>"
-        )
-
-        # Run diagnostics in a background thread
-        from PyQt6.QtCore import QThread, pyqtSignal
-
-        class DiagnosticsWorker(QThread):
-            """Background worker for running diagnostics."""
-
-            finished = pyqtSignal(dict)
-            error = pyqtSignal(str)
-
-            def __init__(self, launcher: Any | None = None) -> None:
-                super().__init__()
-                self._launcher = launcher
-
-            def run(self) -> None:
-                try:
-                    from src.launchers.launcher_diagnostics import LauncherDiagnostics
-
-                    diag = LauncherDiagnostics()
-                    results = diag.run_all_checks()
-
-                    if self._launcher and hasattr(self._launcher, "available_models"):
-                        results["runtime_state"] = {
-                            "available_models_count": len(
-                                self._launcher.available_models
-                            ),
-                            "available_model_ids": list(
-                                self._launcher.available_models.keys()
-                            ),
-                            "model_order_count": len(self._launcher.model_order),
-                            "model_order": self._launcher.model_order,
-                            "model_cards_count": len(self._launcher.model_cards),
-                            "selected_model": self._launcher.selected_model,
-                            "docker_available": self._launcher.docker_available,
-                            "registry_loaded": self._launcher.registry is not None,
-                        }
-
-                    self.finished.emit(results)
-                except ImportError as e:
-                    self.error.emit(str(e))
-
-        self._diag_worker = DiagnosticsWorker(self._launcher)
-        self._diag_worker.finished.connect(self._on_diagnostics_ready)
-        self._diag_worker.error.connect(self._on_diagnostics_error)
-        self._diag_worker.start()
-
-    def _on_diagnostics_ready(self, results: dict[str, Any]) -> None:
-        """Handle completed diagnostics results.
-
-        Args:
-            results: The diagnostics results dictionary.
-        """
-        self._diagnostics_data = results
-        self._render_diagnostics(results)
-
-    def _on_diagnostics_error(self, error_msg: str) -> None:
-        """Handle diagnostics error.
-
-        Args:
-            error_msg: The error message.
-        """
-        self._diag_browser.setHtml(
-            f"<p style='color:#f85149;'>Error running diagnostics: {error_msg}</p>"
-        )
-
-    def _sync_shared_tools(self) -> None:
-        """Run submodule synchronization and fetch latest remote commits in a background thread."""
-        from PyQt6.QtWidgets import QMessageBox
-
-        reply = QMessageBox.question(
-            self,
-            "Sync Shared Tools",
-            "Are you sure you want to synchronize git submodules and fetch latest remote updates?\n\n"
-            "This will run 'git submodule update --init --recursive' in the background.",
-            QMessageBox.StandardButton.Yes | QMessageBox.StandardButton.No,
-            QMessageBox.StandardButton.No,
-        )
-        if reply == QMessageBox.StandardButton.No:
-            return
-
-        self.btn_sync_tools.setEnabled(False)
-        self.btn_sync_tools.setText("Syncing...")
-        self._diag_browser.append(
-            "<p style='color:#58a6ff;'>Starting sync of shared tools...</p>"
-        )
-
-        from PyQt6.QtCore import QThread, pyqtSignal
-        from src.shared.python.data_io.path_utils import get_repo_root
-        from src.launchers.launcher_diagnostics import LauncherDiagnostics
-        import subprocess
-
-        class SubmoduleSyncWorker(QThread):
-            finished = pyqtSignal(bool, str)
-
-            def __init__(self, repos_root: Path) -> None:
-                super().__init__()
-                self.repos_root = repos_root
-
-            def run(self) -> None:
-                try:
-                    # 1. git submodule update
-                    result = subprocess.run(
-                        ["git", "submodule", "update", "--init", "--recursive"],
-                        cwd=str(self.repos_root),
-                        capture_output=True,
-                        text=True,
-                        check=True,
-                        encoding="utf-8",
-                        timeout=60.0,
-                    )
-
-                    # 2. Fetch updates for parent repo
-                    subprocess.run(
-                        ["git", "fetch"],
-                        cwd=str(self.repos_root),
-                        capture_output=True,
-                        timeout=15.0,
-                    )
-
-                    # 3. Fetch updates for submodule
-                    submodule_dir = self.repos_root / "vendor" / "ud-tools"
-                    if submodule_dir.is_dir():
-                        subprocess.run(
-                            ["git", "fetch"],
-                            cwd=str(submodule_dir),
-                            capture_output=True,
-                            timeout=15.0,
-                        )
-
-                    # 4. Fetch updates for sibling Tools
-                    sibling_root = LauncherDiagnostics._find_sibling_tools_root()
-                    if sibling_root:
-                        subprocess.run(
-                            ["git", "fetch"],
-                            cwd=str(sibling_root),
-                            capture_output=True,
-                            timeout=15.0,
-                        )
-
-                    output_msg = (
-                        result.stdout or "Submodules synchronized successfully."
-                    )
-                    self.finished.emit(True, output_msg)
-                except (subprocess.SubprocessError, FileNotFoundError, OSError) as e:
-                    self.finished.emit(False, str(e))
-
-        repos_root = get_repo_root()
-        self._sync_worker = SubmoduleSyncWorker(repos_root)
-        self._sync_worker.finished.connect(self._on_sync_finished)
-        self._sync_worker.start()
-
-    def _on_sync_finished(self, success: bool, output: str) -> None:
-        """Handle completion of the synchronization worker thread."""
-        self.btn_sync_tools.setEnabled(True)
-        self.btn_sync_tools.setText("Sync Shared Tools")
-
-        from PyQt6.QtWidgets import QMessageBox
-
-        if success:
-            QMessageBox.information(
-                self,
-                "Sync Complete",
-                "Shared tools synchronization completed successfully.",
-            )
-            # Re-run diagnostics to update status
-            self._refresh_diagnostics()
-        else:
-            QMessageBox.critical(
-                self,
-                "Sync Failed",
-                f"Failed to synchronize shared tools:\n\n{output}",
-            )
 
     def _compare_versions(self, installed: str, required_spec: str) -> bool:
         """Compare installed version string with required specification (e.g., '>=1.26.4').
@@ -1999,159 +1086,6 @@ class SettingsWidget(QWidget):
                 "<p>Could not access the system clipboard. The command is:</p>"
                 f"<code>{cmd}</code>",
             )
-
-    def _create_processes_tab(self) -> QWidget:
-        """Create the Processes tab for managing running tasks/subprocesses."""
-        tab = QWidget()
-        tab_layout = QVBoxLayout(tab)
-        tab_layout.setContentsMargins(15, 15, 15, 15)
-        tab_layout.setSpacing(10)
-
-        # Status text header
-        self.lbl_processes_status = QLabel("Active Processes")
-        self.lbl_processes_status.setStyleSheet(
-            "font-size: 14px; font-weight: bold; color: #ffffff;"
-        )
-        tab_layout.addWidget(self.lbl_processes_status)
-
-        # Scroll area for the processes list
-        scroll = QScrollArea()
-        scroll.setWidgetResizable(True)
-        scroll.setStyleSheet(
-            "background-color: transparent; border: 1px solid rgba(255, 255, 255, 0.1); border-radius: 8px;"
-        )
-
-        self.processes_container_widget = QWidget()
-        self.processes_container_widget.setStyleSheet("background-color: transparent;")
-        self.processes_layout = QVBoxLayout(self.processes_container_widget)
-        self.processes_layout.setContentsMargins(10, 10, 10, 10)
-        self.processes_layout.setSpacing(6)
-        self.processes_layout.addStretch()  # bottom stretch
-
-        scroll.setWidget(self.processes_container_widget)
-        tab_layout.addWidget(scroll)
-
-        # Bottom row with Kill All button
-        btn_layout = QHBoxLayout()
-        btn_layout.addStretch()
-        self.btn_kill_all_procs = QPushButton("Kill All Processes")
-        self.btn_kill_all_procs.setCursor(Qt.CursorShape.PointingHandCursor)
-        self.btn_kill_all_procs.setStyleSheet("""
-            QPushButton {
-                background-color: rgba(220, 53, 69, 0.2);
-                border: 1px solid rgba(220, 53, 69, 0.4);
-                border-radius: 4px;
-                padding: 6px 14px;
-                color: #ff6b6b;
-                font-weight: bold;
-                font-size: 11px;
-            }
-            QPushButton:hover {
-                background-color: rgba(220, 53, 69, 0.4);
-                color: #ffffff;
-            }
-        """)
-        self.btn_kill_all_procs.clicked.connect(self._on_kill_all_clicked)
-        btn_layout.addWidget(self.btn_kill_all_procs)
-        tab_layout.addLayout(btn_layout)
-
-        return tab
-
-    def _on_kill_all_clicked(self) -> None:
-        if self._launcher and hasattr(self._launcher, "_kill_all_processes"):
-            self._launcher._kill_all_processes()
-            self.refresh_processes_ui()
-
-    def _on_kill_clicked(self, name: str) -> None:
-        if self._launcher and hasattr(self._launcher, "_kill_process_by_name"):
-            self._launcher._kill_process_by_name(name)
-            self.refresh_processes_ui()
-
-    def refresh_processes_ui(self) -> None:
-        """Dynamically populate active processes list row-by-row (GUI thread only)."""
-        # Clear existing rows except the stretch spacer at the end
-        while self.processes_layout.count() > 1:
-            item = self.processes_layout.takeAt(0)
-            if item.widget():
-                item.widget().deleteLater()
-
-        # Check active processes
-        running = []
-        if (
-            self._launcher
-            and hasattr(self._launcher, "running_processes")
-            and hasattr(self._launcher, "process_manager")
-        ):
-            with self._launcher.process_manager._process_lock:
-                for name, proc in list(self._launcher.running_processes.items()):
-                    if proc.poll() is None:
-                        running.append(name)
-
-        if not running:
-            self.lbl_processes_status.setText("No active processes.")
-            self.lbl_processes_status.setStyleSheet(
-                "font-size: 14px; font-weight: bold; color: gray;"
-            )
-            self.btn_kill_all_procs.setEnabled(False)
-            return
-
-        self.lbl_processes_status.setText(f"Active Processes ({len(running)})")
-        self.lbl_processes_status.setStyleSheet(
-            "font-size: 14px; font-weight: bold; color: #ffffff;"
-        )
-        self.btn_kill_all_procs.setEnabled(True)
-
-        for name in running:
-            row = QWidget()
-            row.setObjectName("ProcessRow")
-            row.setStyleSheet("""
-                #ProcessRow {
-                    background-color: rgba(255, 255, 255, 0.03);
-                    border-radius: 6px;
-                    border: 1px solid rgba(255, 255, 255, 0.05);
-                }
-                #ProcessRow:hover {
-                    background-color: rgba(255, 255, 255, 0.06);
-                }
-            """)
-            row_layout = QHBoxLayout(row)
-            row_layout.setContentsMargins(10, 8, 10, 8)
-
-            # Green status dot
-            dot = QLabel("●")
-            dot.setStyleSheet("color: #2ecc71; font-size: 14px; margin-right: 6px;")
-            row_layout.addWidget(dot)
-
-            # Process Name
-            lbl_name = QLabel(name)
-            lbl_name.setStyleSheet("color: #e0e0e0; font-size: 12px; font-weight: 500;")
-            row_layout.addWidget(lbl_name)
-            row_layout.addStretch()
-
-            # Kill button
-            btn_kill = QPushButton("Kill")
-            btn_kill.setCursor(Qt.CursorShape.PointingHandCursor)
-            btn_kill.setStyleSheet("""
-                QPushButton {
-                    background-color: rgba(220, 53, 69, 0.1);
-                    border: 1px solid rgba(220, 53, 69, 0.3);
-                    border-radius: 4px;
-                    padding: 4px 10px;
-                    color: #ff6b6b;
-                    font-size: 11px;
-                }
-                QPushButton:hover {
-                    background-color: rgba(220, 53, 69, 0.3);
-                    color: #ffffff;
-                }
-            """)
-            btn_kill.clicked.connect(
-                lambda checked=False, n=name: self._on_kill_clicked(n)
-            )
-            row_layout.addWidget(btn_kill)
-
-            # Insert at the top (before the spacer at the end)
-            self.processes_layout.insertWidget(self.processes_layout.count() - 1, row)
 
 
 class SettingsDialog(QDialog):
